Move term to separate module
