+++ src/clinlp/ie/__init__.py
@@ -1,3 +1,4 @@
-from .entity import ENTS_KEYWORD, RuleBasedEntityMatcher, Term, create_concept_dict
+from .entity import SPANS_KEY, RuleBasedEntityMatcher, create_concept_dict
+from .term import Term
 
-__all__ = ["ENTS_KEYWORD", "RuleBasedEntityMatcher", "Term", "create_concept_dict"]
+__all__ = ["SPANS_KEY", "RuleBasedEntityMatcher", "Term", "create_concept_dict"]
