+++ src/clinlp/ie/entity.py
@@ -8,6 +8,7 @@ from spacy.language import Doc, Language
 from spacy.matcher import Matcher, PhraseMatcher
 from spacy.tokens import Span
 
+from clinlp.ie.term import Term
 from clinlp.util import clinlp_autocomponent
 
 SPANS_KEY = "ents"
@@ -21,46 +22,6 @@ _defaults_clinlp_ner = {
     "pseudo": False,
 }
 _non_phrase_matcher_fields = ["proximity", "fuzzy", "fuzzy_min_len"]
-
-
-class Term(pydantic.BaseModel):
-    phrase: str
-    attr: Optional[str] = None
-    proximity: Optional[int] = None
-    fuzzy: Optional[int] = None
-    fuzzy_min_len: Optional[int] = None
-    pseudo: Optional[bool] = None
-
-    model_config = {"extra": "ignore"}
-
-    def __init__(self, phrase: str, **kwargs):
-        """This init makes sure Term accepts phrase as a positional argument,
-        which is more readable in large concept lists."""
-        super().__init__(phrase=phrase, **kwargs)
-
-    def to_spacy_pattern(self, nlp: Language):
-        fields = {
-            field: getattr(self, field) or _defaults_clinlp_ner[field]
-            for field in ["attr", "proximity", "fuzzy", "fuzzy_min_len", "pseudo"]
-        }
-
-        spacy_pattern = []
-
-        phrase_tokens = [token.text for token in nlp.tokenizer(self.phrase)]
-
-        for i, token in enumerate(phrase_tokens):
-            if (fields["fuzzy"] > 0) and (len(token) >= fields["fuzzy_min_len"]):
-                token_pattern = {f"FUZZY{fields['fuzzy']}": token}
-            else:
-                token_pattern = token
-
-            spacy_pattern.append({fields["attr"]: token_pattern})
-
-            if i != len(phrase_tokens) - 1:
-                for _ in range(fields["proximity"]):
-                    spacy_pattern.append({"OP": "?"})
-
-        return spacy_pattern
 
 
 def create_concept_dict(path: str, concept_col: str = "concept") -> dict:
